Add constants class and use correct input values

+++ TruckPad/app/src/main/java/com/truckpad/androidcase/model/RouteResponse.kt
@@ -4,9 +4,9 @@ import com.google.gson.annotations.SerializedName
 
 data class RouteResponse(
     @SerializedName("cached") val cached: Boolean,
-    @SerializedName("distance") val distance: Int,
+    @SerializedName("distance") val distance: Double,
     @SerializedName("distance_unit") val distanceUnit: String,
-    @SerializedName("duration") val duration: Int,
+    @SerializedName("duration") val duration: Double,
     @SerializedName("duration_unit") val durationUnit: String,
     @SerializedName("fuel_cost") val fuelCost: Double,
     @SerializedName("fuel_cost_unit") val fuelCostUnit: String,
@@ -16,7 +16,7 @@ data class RouteResponse(
     @SerializedName("points") val points: List<Point>,
     @SerializedName("provider") val provider: String,
     @SerializedName("route") val route: List<List<List<Double>>>,
-    @SerializedName("toll_cost") val tollCost: Int,
+    @SerializedName("toll_cost") val tollCost: Double,
     @SerializedName("toll_cost_unit") val tollCostUnit: String,
     @SerializedName("toll_count") val tollCount: Int,
     @SerializedName("total_cost") val totalCost: Double
